HSEARCH-4175 Merge PojoValueBridgeToDocumentValueConverter with PojoValueBridgeFromDocumentValueConverter
